requirements.txt
attrs==22.2.0
aws-cdk-lib==2.110.1
aws-cdk.asset-awscli-v1==2.2.201
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v5==2.0.78
aws-cdk.asset-node-proxy-agent-v6==2.0.1
cattrs==22.2.0
cdk-nag==2.27.102
constructs==10.1.271
exceptiongroup==1.1.0
importlib-resources==5.12.0
jsii==1.92.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.5.0
aws_solutions_constructs.aws_kinesis_streams_kinesis_firehose_s3==2.46.0